@@ -49,7 +49,7 @@ class _MyAppState extends State<MyApp> {
     return Scaffold(
       appBar: AppBar(title: const Text('Plugin example app')),
       body: Center(
-        child: RaisedButton(
+        child: ElevatedButton(
           child: Text("Open PDF"),
           onPressed: () => Navigator.push(
             context,
@@ -62,7 +62,7 @@ class _MyAppState extends State<MyApp> {
 }
 
 class PDFScreen extends StatelessWidget {
-  String pathPDF = "";
+  final String pathPDF;
   PDFScreen(this.pathPDF);
 
   @override
